Update frontend services and add backend CORS
This adds functionality to the front end service to maintain a list of
questions retrieved from the backend, which has been slightly altered to
include a CORS package. In addition, there is a dummy component to hold
all questions available from the backend, though this should be removed
from the homepage once testing is complete.